Add TSGI Layer Between Router and Server (#81)
Changes done:
1. decompose the code into modular pieces
2. decouple router and server parts
3. introduce nginx as possible Web Server via TSGI adapter
4. enable router tests on NGINX

Note: this change breaks connection stealing.
