+++ test/Procfile.test.nginx
@@ -1,0 +1,2 @@
+nginx: nginx -g "daemon off;" -c $PWD/test/nginx.conf
+nginx_tsgi_test: SERVER_TYPE=nginx ./test/http.test.lua 2>&1
